change password page completed

+++ AGSWeb/ags-web/pages/api/identity/users/[userId].js
@@ -4,9 +4,14 @@ import UsersHelper from '../../../../helpers/Identity/api/usersHelper.js'
 export default async function UsersHandler(req, res) {
     const usersHelper = new UsersHelper(req,res);
     const { method, query } = req;
-    
     const { userId } = query
 
+    if (method == "POST" && userId.toLowerCase() == "changepw"){
+        const result = await usersHelper.ChangePassword(req.body.changePWReuqest);
+        res.setHeader('Content-Type', 'application/json');
+        res.end(JSON.stringify(result));
+        return ;
+    }
     
     if (method == "DELETE"){
         const result = await usersHelper.DeleteUser(userId);
